third_party/WebKit/LayoutTests/fast/harness/internals-observe-gc.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<body>
<script>
description('Tests that the window.internals.observeGC hook works.');

shouldBe('typeof window.internals.observeGC', '"function"',
         'this test requires window.internals');

// "Generic Kid's Movie III": ... where nobody dies.

var valueA = {};
// Do not pass the object directly to observeGC function. This may
// remain live on this function's stack preventing GC from collecting
// it. Accessing the object inside an inner function will prevent any
// unneeded references on this function's stack.
observationA = internals.observeGC((() => {return valueA;})());
gc();
shouldBeFalse('observationA.wasCollected');
// value ineligible for GC should not be flagged as collected
valueA = null;
observationA = null;

// "Romeo and GCuliet": Romeo JavaScript finds G.uliet C.apulet and dies.

var valueB = {};
// Do not pass the object directly to observeGC function. This may
// remain live on this function's stack preventing GC from collecting
// it. Accessing the object inside an inner function will prevent any
// unneeded references on this function's stack.
observationB = internals.observeGC((() => {return valueB;})());
valueB = null;
gc();
shouldBeTrue('observationB.wasCollected');
// value eligible for GC should be flagged as collected
observationB = null;

// "One DOM Tree Hill": A family struggles with the loss of
// innocence. And a DOM node.

var valueC = document.createElement('div');
// Do not pass the object directly to observeGC function. This may
// remain live on this function's stack preventing GC from collecting
// it. Accessing the object inside an inner function will prevent any
// unneeded references on this function's stack.
observationC = internals.observeGC((() => {return valueC;})());
valueC = null;
gc();
shouldBeTrue('observationC.wasCollected');
// DOM node eligible for GC should be flagged as collected
observationC = null;

// Now, movies that failed:

shouldThrow('internals.observeGC(undefined)',
    '"TypeError: value to observe is null or undefined"');
shouldThrow('internals.observeGC(null)',
   '"TypeError: value to observe is null or undefined"');

// Try to create objects and observers that will die at once

var valueD = {};
// Do not pass the object directly to observeGC function. This may
// remain live on this function's stack preventing GC from collecting
// it. Accessing the object inside an inner function will prevent any
// unneeded references on this function's stack.
var observerD = internals.observeGC((() => {return valueD;})());
valueD.observer = observerD;
observerD.observed = valueD;
valueD = observerD = null;
gc();
testPassed('did not crash');

var successfullyParsed = true;
</script>
